requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0

# Database and storage
supabase==2.22.0
psycopg2-binary==2.9.5

# AI and ML
boto3==1.34.0
openai==1.6.0
numpy==1.24.0

# Google APIs
google-api-python-client==2.110.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Caching (optional)
redis==5.0.1
aioredis==2.0.1

# Email utilities
email-validator>=2.0.0

# JSON and data processing
orjson==3.9.10
pandas==2.1.4

# Background tasks (optional)
celery==5.3.4
kombu==5.3.4

# Monitoring and metrics
prometheus-client==0.19.0

# Configuration management
dynaconf==3.2.4

# Rate limiting (optional)
# slowapi==0.1.9

# CORS handled by FastAPI middleware

# File handling
aiofiles==23.2.1

# Validation and serialization
marshmallow==3.20.1
cerberus==1.3.5

# Time and scheduling
schedule==1.2.0
croniter==2.0.1

# Vector operations
scikit-learn==1.3.2
scipy==1.11.4

# Development tools
pre-commit==3.6.0
bandit==1.7.5
safety==2.3.5

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8